Resolved various conflicts with test runners that blocked using routing.
These changes bring the project's test running in line with the recommendations
made in the Meteor Guide and as illustrated in the Meteor Guide's test
application.

`meteor npm test` is now a console test runner that runs all tests once and
finishes, much like the former `meteor npm run test-cli-once`. The CircleCI
config was updated to reflect the change.

`meteor npm test-watch` is now a web-based test runner that watches for code
changes and re-runs as needed, which matches what was previous used for `meteor
npm test`.

Meteor and all packages were also updated to bring in a bug fix for an issue
that previously blocked using dispatch:mocha-phantomjs and
practicalmeteor:mocha together.